Update index.html
I moved the Thunderbolt 5 copy to after Thunderbolt 3 + 4, then moved the Connectors and Ports to the end of the section. I also moved the line "Each iteration of Thunderbolt is backwards compatible with previous versions" and the note on Thunderbolt 4 to the main paragraph at the top. Lastly I changed the date introduced from 2013 to 2023